% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ar_change.R
\name{fit_ar_change}
\alias{fit_ar_change}
\title{Fit AR1 or AR2 changepoint model with optional trend}
\usage{
fit_ar_change(x, ar_order = 1, trend = FALSE, method = "PELT", minseglen = 5)
}
\arguments{
\item{x}{Numeric vector.}

\item{ar_order}{Either 1 or 2.}

\item{trend}{Logical. If TRUE, include a time trend.}

\item{method}{Changepoint method, either "PELT" or "AMOC".}

\item{minseglen}{Positive minimum segment length.}
}
\value{
A fitted `cpt.reg` object.
}
\description{
Fit AR1 or AR2 changepoint model with optional trend
}
